requirements.txt
flask==2.3.2
flask-cors==3.0.10
flask-limiter==2.8.1
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
google==3.0.0